+++ include/grabber/dda/DDAWrapper.h
@@ -6,18 +6,44 @@
 class DDAWrapper : public GrabberWrapper
 {
 public:
-	static constexpr const char *GRABBERTYPE = "DDA";
+	static constexpr const char* GRABBERTYPE = "DDA";
 
-	DDAWrapper(int updateRate_Hz = GrabberWrapper::DEFAULT_RATE_HZ, int display = 0,
-	           int pixelDecimation = GrabberWrapper::DEFAULT_PIXELDECIMATION, int cropLeft = 0, int cropRight = 0,
-	           int cropTop = 0, int cropBottom = 0);
+	///
+	/// Constructs the DDA frame grabber with a specified grab size and update rate.
+	///
+	/// @param[in] updateRate_Hz     The image grab rate [Hz]
+	/// @param[in] display           Display to be grabbed
+	/// @param[in] pixelDecimation   Decimation factor for image [pixels]
+	/// @param[in] cropLeft          Remove from left [pixels]
+	/// @param[in] cropRight 	     Remove from right [pixels]
+	/// @param[in] cropTop           Remove from top [pixels]
+	/// @param[in] cropBottom        Remove from bottom [pixels]
+	///
+	explicit DDAWrapper(int updateRate_Hz = GrabberWrapper::DEFAULT_RATE_HZ,
+		int display = 0,
+		int pixelDecimation = GrabberWrapper::DEFAULT_PIXELDECIMATION,
+		int cropLeft = 0, int cropRight = 0,
+		int cropTop = 0, int cropBottom = 0
+	);
 
-	DDAWrapper(const QJsonDocument &grabberConfig = QJsonDocument());
+	///
+	/// Constructs the DDA frame grabber from configuration settings
+	///
+	explicit DDAWrapper(const QJsonDocument& grabberConfig = QJsonDocument());
 
-	virtual ~DDAWrapper(){};
+	///
+	/// Starts the grabber, if available
+	///
+	bool start() override;
 
 public slots:
-	virtual void action();
+
+	///
+	/// Performs a single frame grab and computes the LED-colors
+	///
+	void action();
+
+	void handleSettingsUpdate(settings::type type, const QJsonDocument& grabberConfig) override;
 
 private:
 	DDAGrabber _grabber;