+++ libsrc/hyperion/GrabberWrapper.cpp
@@ -25,9 +25,9 @@ bool GrabberWrapper::GLOBAL_GRABBER_V4L_ENABLE = false;
 bool GrabberWrapper::GLOBAL_GRABBER_AUDIO_ENABLE = false;
 
 GrabberWrapper::GrabberWrapper(const QString& grabberName, Grabber * ggrabber, int updateRate_Hz)
-	: _ggrabber(ggrabber)
+	: _log(Logger::getInstance(("Grabber-" + grabberName).toUpper()))
+	, _ggrabber(ggrabber)
 	, _grabberName(grabberName)
-	, _log(Logger::getInstance(("Grabber-" + grabberName).toUpper()))
 	, _timer(nullptr)
 	, _updateInterval_ms(1000/updateRate_Hz)
 {