+++ libsrc/leddevice/dev_net/LedDeviceUdpDdp.h
@@ -50,7 +50,6 @@ protected:
 	///
 	int open() override;
 
-
 	///
 	/// @brief Writes the RGB-Color values to the LEDs.
 	///