+++ libsrc/leddevice/dev_net/LedDeviceUdpRaw.h
@@ -57,7 +57,6 @@ protected:
 	///
 	int open() override;
 
-
 	///
 	/// @brief Writes the RGB-Color values to the LEDs.
 	///